- move details for "HDMI Audio Passthrough Capabilities" from the end of the device report, after the "Device codec decoder" content. - Add dedicated file and class for the passthrough function into the Codec directory
