+++ app/src/main/java/org/jellyfin/androidtv/util/profile/deviceProfile.kt
@@ -1,14 +1,7 @@
 package org.jellyfin.androidtv.util.profile
 
 import android.content.Context
-import android.media.AudioFormat
-import androidx.annotation.OptIn
-import androidx.media3.common.AudioAttributes
-import androidx.media3.common.C
-import androidx.media3.common.Format
 import androidx.media3.common.MimeTypes
-import androidx.media3.common.util.UnstableApi
-import androidx.media3.exoplayer.audio.AudioCapabilities
 import org.jellyfin.androidtv.constant.Codec
 import org.jellyfin.androidtv.preference.UserPreferences
 import org.jellyfin.androidtv.preference.constant.AudioBehavior
@@ -567,30 +560,4 @@ private fun DeviceProfileBuilder.subtitleProfile(
 	if (external) subtitleProfile(format, SubtitleDeliveryMethod.EXTERNAL)
 	if (hls) subtitleProfile(format, SubtitleDeliveryMethod.HLS)
 	if (encode) subtitleProfile(format, SubtitleDeliveryMethod.ENCODE)
-}
-
-@OptIn(UnstableApi::class)
-fun isPassthroughAudioAvailable(context: Context, mimetype: String): Boolean {
-	// Def audio attributes
-	val audioAttributes = AudioAttributes.Builder()
-		.setUsage(C.USAGE_MEDIA)
-		.setContentType(C.AUDIO_CONTENT_TYPE_MOVIE)
-		.build()
-	// Get audio capabilities
-	val audioCapabilities = AudioCapabilities.getCapabilities(
-		context,
-		audioAttributes,
-		null,
-		listOf(
-			AudioFormat.CHANNEL_OUT_STEREO,
-			AudioFormat.CHANNEL_OUT_5POINT1 )
-	)
-	// Set audio format for a passthrough audio codec 2.0 check
-	val format = Format.Builder()
-		.setSampleMimeType(mimetype)
-		.setChannelCount(Integer.bitCount(AudioFormat.CHANNEL_OUT_STEREO))
-		.setSampleRate(Format.NO_VALUE)
-		.build()
-	// Test Passthrough Direct Playback
-	return audioCapabilities.isPassthroughPlaybackSupported(format, audioAttributes)
 }
